+++ app/components/omnibox/services/search-service.js
@@ -58,6 +58,9 @@ angular.module('omnibox')
     this.search = function (searchString, state) {
 
       var getSearch = function (params) {
+
+        var MINIMUM_SEARCH_QUERY_LENGTH = 2;
+
         // Cancel consecutive calls.
         if (localPromise.resolve) {
           localPromise.resolve({data: {results: []}});
@@ -65,15 +68,23 @@ angular.module('omnibox')
 
         localPromise = $q.defer();
 
-        return $http({
-          url: 'api/v2/search/',
-          method: 'GET',
-          params: params,
-          timeout: localPromise.promise
-        })
-        .then(function (response) {
-          return response.data;
-        }, errorFn);
+        // Only send request if searchstring is longer than 2.
+        // Otherwise return zero relevant searches.
+        if (params.q.length > MINIMUM_SEARCH_QUERY_LENGTH) {
+          return $http({
+            url: 'api/v2/search/',
+            method: 'GET',
+            params: params,
+            timeout: localPromise.promise
+          })
+          .then(function (response) {
+            return response.data;
+          }, errorFn);
+        }
+        else {
+          localPromise.resolve({results: []});
+          return localPromise.promise;
+        }
       };
 
       var errorFn = function (err) {