+++ CHANGES.rst
@@ -3,7 +3,8 @@ Changelog of lizard-nxt client
 
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
--
+
+- Fix attempting to draw assets (filters) without geometry.
 
 
 Release 3.0.11 (2016-4-26)