+++ CHANGES.rst
@@ -12,6 +12,8 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 
 - Remove old graph code (donut).
 
+- Fix showing rain and other data in omnibox.
+
 
 Release 3.0.11 (2016-4-26)
 ---------------------