+++ CHANGES.rst
@@ -16,6 +16,9 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 
 - Fix search not cleaning scope after enter keypress.
 
+- Fix infinite digest loop, copy land use data.
+
+- Fix landuse label falling outside of svg.
 
 Release 3.0.11 (2016-4-26)
 ---------------------