+++ CHANGES.rst
@@ -20,6 +20,8 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 
 - Fix landuse label falling outside of svg.
 
+- Fix clicking on cluster zooms too far in.
+
 Release 3.0.11 (2016-4-26)
 ---------------------