+++ CHANGES.rst
@@ -22,6 +22,8 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 
 - Fix clicking on cluster zooms too far in.
 
+- Fix multiplying annotations when button bashing.
+
 Release 3.0.11 (2016-4-26)
 ---------------------